Add new command line flag to keep generated projects
Sometimes is important to keep the baked projects in order
to do some debugging. This commit add the feature, tests
and documentation. The idea is that someone could do:

    pytest -k a_hard_to_debug_test --keep-cookies-directories

In this scenario, the tmp directories will not be removed
and the user will be able to access them. This won't clutter
the tmp directory because pytest will remove entries older than 3
temporary directories anyways, see:

https://docs.pytest.org/en/latest/tmpdir.html#the-default-base-temporary-directory